WiP: Improve duplicated OID
